requirements.txt
south
django-compressor
Mezzanine==3.1.4
icalendar
